T-050/T-051: Extract AnalyticsTabFragment + HealthTabFragment from HistoryFragment
- Create AnalyticsTabFragment.kt with fragment_analytics_tab.xml: moves all
  timeline, hero stats, bar chart, weekly comparison, streaks, period toggle,
  and expandable deep-dive sections out of HistoryFragment
- Create HealthTabFragment.kt with fragment_health_tab.xml: moves the Health
  & Intake card (total grams / weekly grams / session split)
- HistoryFragment.kt reduced to a stub with TODO for T-052 tab host wiring
